refactor: give the bounded member read a home both readers can use (#773)
`read_bounded_member` lived in the resource endpoint's query adapter, and
the publication parser needs the same read for the documents it resolves a
book's structure from. What it knows is about zip members rather than
about publications -- that a declared size bounds the result of
`ZipFile.read()` and not its work -- so it moves to
`infrastructure/common/zip_members.py`, re-exported where it was, leaving
the resource endpoint's monkeypatch and its own tests untouched.
