+++ backend/src/infrastructure/web_reader/queries/publication_resource_query.py
@@ -20,7 +20,15 @@ from src.application.web_reader.queries.publication_resource import (
 from src.domain.common.value_objects.ids import BookId, UserId
 from src.domain.library.exceptions import InvalidEbookError
 from src.domain.web_reader.exceptions import PublicationResourceNotFoundError
+from src.infrastructure.common.zip_members import read_bounded_member
 from src.infrastructure.web_reader.queries.stored_epub import PublicationQuery, StoredEpub
+
+__all__ = [
+    "MAX_RESOURCE_BYTES",
+    "PublicationResourceQuery",
+    "check_member_is_servable",
+    "read_bounded_member",
+]
 
 # Long enough that two files of one book cannot collide by accident, short
 # enough to stay readable in a log line or a browser's network panel.
@@ -185,34 +193,6 @@ def check_member_is_servable(entry: zipfile.ZipInfo) -> None:
             f"compressed stream of {entry.compress_size}",
             "epub",
         )
-
-
-def read_bounded_member(archive: zipfile.ZipFile, entry: zipfile.ZipInfo) -> bytes:
-    """Read one member without letting it decide how much memory that takes.
-
-    ``ZipFile.read()`` returns no more than the declared size but does not
-    *work* within it: measured on CPython 3.13, reading a member that declares
-    10 bytes and really inflates to 200 MB returns 10 bytes after allocating
-    437 MiB, because the decompressor is handed an unbounded output limit and
-    only the result is truncated. ``open(entry).read(n)`` passes ``n`` down as
-    that limit, so asking for no more than the declaration bounds the work as
-    well as the answer -- the same member then costs 53 KiB.
-
-    Call :func:`check_member_is_servable` first; this trusts the declaration to
-    be one worth reading up to.
-
-    A member that inflates past what it declared fails its CRC-32, which
-    ``zipfile`` raises as ``BadZipFile``; that is a broken publication rather
-    than a server fault, so it reads as one.
-
-    Raises:
-        InvalidEbookError: If the member is not readable.
-    """
-    try:
-        with archive.open(entry) as member:
-            return member.read(entry.file_size + 1)
-    except (OSError, zipfile.BadZipFile, EOFError) as e:
-        raise InvalidEbookError(f"resource {entry.filename!r} cannot be read: {e!s}", "epub") from e
 
 
 def _version(stored: StoredEpub, path: str) -> str:
